requirements.txt
numpy
neptune-client
psutil
matplotlib
scipy
black
pylint
isort
sinkhorn_knopp
pandas
seaborn